requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
defusedxml==0.7.1
Django==3.2.7
django-environ==0.7.0
djangorestframework==3.12.4
djangorestframework-xml==2.0.0
googlemaps==4.5.3
idna==3.2
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
